[Feat] TextBox 내부 Timer 구현 (#43)
* feat: 지원하기 input에 type 추가

* init: Timer 작업 파일 생성

* fix: Timer 컴포넌트 Input 파일 내부로

* feat: Timer 기능 구현

* feat: Timer 시작, 재시작 기능 구현

* chore: Input 파일에 주석 추가

* feat: InputButton 서브컴포넌트로 분리

* fix: 코드리뷰 반영

* move: TextBox 합성 컴포넌트 폴더 분리
