+++ src/common/components/Input/Description/index.tsx
@@ -1,0 +1,9 @@
+import { descriptionVar } from './style.css';
+import { TextBoxProps } from '../types';
+
+// TextBox 내부 Input 하단의 부가텍스트
+const Description = ({ children, styleType = 'default' }: Pick<TextBoxProps, 'children' | 'styleType'>) => (
+  <div className={descriptionVar[styleType]}>{children}</div>
+);
+
+export default Description;
