+++ src/common/components/Input/Timer/utils/formatTimer.ts
@@ -1,0 +1,8 @@
+const formatTimer = (seconds: number) => {
+  const min = Math.floor(seconds / 60);
+  const sec = seconds % 60;
+
+  return `${min}:${sec < 10 ? '0' + sec : sec}`;
+};
+
+export default formatTimer;
