+++ src/common/components/Input/index.tsx
@@ -1,80 +1,5 @@
-/* eslint-disable indent */
-import { createContext, useContext } from 'react';
-
-import { circle, title, inputLine, containerVar, inputVar, descriptionVar } from './style.css';
-import { TextBoxProps } from './types';
-
-const FormContext = createContext({} as Pick<TextBoxProps, 'required' | 'formObject'>);
-
-export const InputLine = ({
-  label,
-  pattern,
-  validate,
-  button,
-  errorText,
-  ...inputElementProps
-}: Omit<TextBoxProps, 'size' | 'children' | 'formObject'>) => {
-  const {
-    required,
-    formObject: { register, formState, clearErrors, trigger },
-  } = useContext(FormContext);
-  const { errors } = formState;
-
-  return (
-    <>
-      <div className={inputLine}>
-        <input
-          id={label}
-          className={inputVar[errors?.[label] ? 'error' : 'default']}
-          {...inputElementProps}
-          {...register(label, {
-            required: required && '필수 입력 항목이에요',
-            pattern:
-              pattern && errorText
-                ? {
-                    value: pattern,
-                    message: errorText,
-                  }
-                : undefined,
-            validate: validate,
-            onBlur: (e) => {
-              if (!pattern || e.currentTarget.value === '') return;
-              trigger(label);
-            },
-            onChange: () => clearErrors && clearErrors(label),
-          })}
-        />
-        {button}
-      </div>
-      {errors[label] && (
-        <Description styleType="error">
-          <p>{errors[label]?.message as string}</p>
-        </Description>
-      )}
-    </>
-  );
-};
-
-export const Description = ({ children, styleType = 'default' }: Pick<TextBoxProps, 'children' | 'styleType'>) => (
-  <div className={descriptionVar[styleType]}>{children}</div>
-);
-
-export const TextBox = ({
-  children,
-  label,
-  size = 'md',
-  required,
-  formObject,
-}: Pick<TextBoxProps, 'children' | 'label' | 'size' | 'required' | 'formObject'>) => {
-  return (
-    <FormContext.Provider value={{ required, formObject }}>
-      <div className={containerVar[size]}>
-        <label className={title} htmlFor={label}>
-          <span>{label}</span>
-          {required && <i className={circle} />}
-        </label>
-        {children}
-      </div>
-    </FormContext.Provider>
-  );
-};
+export { default as Description } from './Description';
+export { default as InputButton } from './InputButton';
+export { default as InputLine } from './InputLine';
+export { TextBox } from './TextBox';
+export { default as Timer } from './Timer';
