+++ src/common/components/Input/types.ts
@@ -8,10 +8,19 @@ export interface TextBoxProps extends Omit<InputHTMLAttributes<HTMLInputElement>
   label: string;
   size?: 'sm' | 'md' | 'lg';
   errorText?: string;
-  button?: ReactNode;
   pattern?: RegExp;
   validate?: Validate<FieldValues, TFormValues>;
   formObject: Pick<UseFormReturn, 'register' | 'formState' | 'clearErrors' | 'trigger'>;
-  children: ReactNode;
+  children?: ReactNode;
   styleType?: 'default' | 'error';
+}
+
+export interface InputButtonProps {
+  text: string;
+  disabled: boolean;
+  onClick: () => void;
+}
+export interface TimerProps {
+  isActive: boolean;
+  onResetTimer: () => void;
 }
