+++ src/views/ApplyPage/index.tsx
@@ -1,6 +1,7 @@
+import { useState } from 'react';
 import { useForm, SubmitHandler } from 'react-hook-form';
 
-import { Description, InputLine, TextBox } from '@components/Input';
+import { Description, InputButton, InputLine, TextBox, Timer } from '@components/Input';
 import SelectBox from '@components/Select';
 import { TFormValues, defaultValues } from '@constants/defaultValues';
 
@@ -12,23 +13,14 @@ const MoreButton = () => {
   );
 };
 
-const CheckButton = () => {
-  return (
-    <button
-      type="button"
-      style={{ cursor: 'pointer', backgroundColor: 'green', padding: 10 }}
-      onClick={() => console.log('버튼클릭')}>
-      인증하기
-    </button>
-  );
-};
-
 const ApplyPage = () => {
   const { handleSubmit, ...formObject } = useForm({
     defaultValues: defaultValues,
   });
 
   const onSubmit: SubmitHandler<TFormValues> = (data) => console.log(data);
+
+  const [isActive, setIsActive] = useState(false);
 
   return (
     <form onSubmit={handleSubmit(onSubmit)} style={{ padding: 50 }}>
@@ -39,11 +31,19 @@ const ApplyPage = () => {
       <TextBox label="이메일" required formObject={formObject}>
         <InputLine
           label="이메일"
-          button={<CheckButton />}
           placeholder="이메일을 입력하세요"
           pattern={/^[0-9]*$/}
           errorText="숫자만 입력해주세요"
-        />
+          type="email">
+          <InputButton
+            text="인증하기"
+            disabled={isActive}
+            onClick={() => {
+              setIsActive(true);
+            }}
+          />
+          <Timer isActive={isActive} onResetTimer={() => setIsActive(false)} />
+        </InputLine>
         <InputLine
           label="인증번호"
           placeholder="인증번호를 입력하세요"
