+++ src/views/MainPage/index.tsx
@@ -26,10 +26,10 @@ const MainPage = () => {
         </div>
       </Callout>
       <TextBox label="이메일" formObject={formObject} required>
-        <InputLine label="이메일" placeholder="이메일을 입력해주세요" />
+        <InputLine label="이메일" placeholder="이메일을 입력해주세요" type="email" />
       </TextBox>
       <TextBox label="비밀번호" formObject={formObject} required>
-        <InputLine label="비밀번호" placeholder="비밀번호를 입력해주세요" />
+        <InputLine label="비밀번호" placeholder="비밀번호를 입력해주세요" type="password" />
         <Description>
           <p>비밀번호를 잃어버리셨나요?</p>
           <Link to="/password">비밀번호 재설정하기</Link>
